# gen/server/go/model_actor.go
/*
 * Хранилище фильмов
 *
 * API для управления хранилищем
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi




type Actor struct {

	// Уникальный идентификатор актёра
	Id string `json:"id,omitempty"`

	Name string `json:"name"`

	Gender string `json:"gender"`

	BirthDate string `json:"birthDate"`
}

// AssertActorRequired checks if the required fields are not zero-ed
func AssertActorRequired(obj Actor) error {
	elements := map[string]interface{}{
		"name": obj.Name,
		"gender": obj.Gender,
		"birthDate": obj.BirthDate,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertActorConstraints checks if the values respects the defined constraints
func AssertActorConstraints(obj Actor) error {
	return nil
}