# gen/server/go/model_user.go
/*
 * Хранилище фильмов
 *
 * API для управления хранилищем
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi


type RegisterDTO struct {

	Username string `json:"username"`

	Password string `json:"password"`

	Admin bool `json:"admin,omitempty"`
}



type User struct {

	Id int64 `json:"id,omitempty"`

	Role string `json:"role"`
}

// AssertUserRequired checks if the required fields are not zero-ed
func AssertUserRequired(obj User) error {
	elements := map[string]interface{}{
		"role": obj.Role,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertUserConstraints checks if the values respects the defined constraints
func AssertUserConstraints(obj User) error {
	return nil
}



// AssertRegisterPostRequestRequired checks if the required fields are not zero-ed
func AssertRegisterPostRequestRequired(obj RegisterDTO) error {
	elements := map[string]interface{}{
		"username": obj.Username,
		"password": obj.Password,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertRegisterPostRequestConstraints checks if the values respects the defined constraints
func AssertRegisterPostRequestConstraints(obj RegisterDTO) error {
	return nil
}